refactor: :recycle: isolate pkg transform to pass cypress
