+++ examples/nuxt-app/app.config.ts
@@ -1,5 +1,5 @@
-import pkg from './package.json'
-import { getDpcPkgs } from '@dpc-sdp/ripple-tide-api'
+import pkg from './package.json' assert { type: 'json' }
+import { getDpcPkgs } from '@dpc-sdp/ripple-tide-api/utils'
 
 export default defineAppConfig({
   project: {
