+++ packages/ripple-tide-api/src/utils.ts
@@ -1,0 +1,1 @@
+export { getDpcPkgs } from './utils/index.js'
